♻️ refactor(player): group compound machinery under Interface/compound/
The Interface root mixed five helpers with the index file and four
child component dirs (Information, Controller, PlayListEmpty,
CustomComponent). Move the compound-slot machinery
(compoundSlotMetaMap, useCompoundSlots, useDuplicateSlotWarning)
into a dedicated compound/ subdir with a barrel; rename
compoundSlotMetaMap.ts -> slotMetaMap.ts since "compound" now lives
in the path.

playListPortalContext.ts and playListEmptyContext.ts stay at the
Interface root — they are Interface-level plumbing distinct from the
compound dispatch logic, and Interface/index.tsx provides them.
