+++ package/src/components/AudioPlayer/Interface/compound/index.ts
@@ -1,0 +1,3 @@
+export * from "./slotMetaMap";
+export * from "./useCompoundSlots";
+export * from "./useDuplicateSlotWarning";
